frame-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
pydantic==2.5.0
httpx==0.25.2
av==13.1.0

# Face enhancement dependencies
torch==2.0.1
torchvision==0.15.2
basicsr==1.4.2
facexlib==0.3.0
realesrgan==0.3.0
gfpgan==1.3.8

# CodeFormer dependencies
codeformer-pip==0.0.4
lpips
gdown
